requirements.txt
-i https://pypi.org/simple
appdirs==1.4.4
certifi==2020.6.20
chardet==3.0.4
click==7.1.2; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
cmake==3.18.2.post1
distlib==0.3.1
dlib==19.21.0
face-recognition-models==0.3.0
face-recognition==1.3.0
filelock==3.0.12
idna==2.10; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
kafka-python==2.0.2
numpy==1.19.2; python_version >= '3.6'
opencv-python==4.4.0.44
pillow==8.0.1; python_version >= '3.6'
pyyaml==5.3.1
requests==2.24.0
six==1.15.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
urllib3==1.25.11; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4' and python_version < '4'
virtualenv==20.1.0
